+++ Cargo.toml
@@ -54,6 +54,7 @@ output-name = "rupa"
 site-root = "target/site"
 site-pkg-dir = "pkg"
 style-file = "style/main.scss"
+hash-files = true
 assets-dir = "public"
 site-addr = "127.0.0.1:3000"
 reload-port = 3001
