+++ src/components/mod.rs
@@ -1,3 +1,3 @@
-pub mod errors;
+pub mod error;
 pub mod navigation;
 pub mod dashboard;
